# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prasso Invenbin API",
        "description": "API endpoints for Prasso Invenbin",
        "contact": {
            "email": "bcp@faxt.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a list of categories records with eager loading of products relationship",
                "operationId": "b1b247b2b491b0491c3c728dc33f23f0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "2"
                                            },
                                            "category_name": {
                                                "type": "string",
                                                "example": "Cat A"
                                            },
                                            "parent": {
                                                "type": "string",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category",
                "operationId": "c86cd7a85ad06a4fe905bfab62146952",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific category",
                "description": "Display the specified resource.",
                "operationId": "486849467f411ec5e4bdf111f50ede87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryInputOutput"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "description": "Update a category",
                "operationId": "b648463d94e0caf7b8d842a1008789b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Remove the specified resource from storage.",
                "operationId": "491dfe5b6b58b480181690ede6550e44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryInputOutput": {
                "title": "Category",
                "description": "Erp Category model",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer"
                    },
                    "category_name": {
                        "description": "Name of the category",
                        "type": "string"
                    },
                    "image_file": {
                        "description": "Image file of the category",
                        "type": "string"
                    },
                    "parent_id": {
                        "description": "ID of the parent category",
                        "type": "integer"
                    },
                    "short_description": {
                        "description": "Short description of the category",
                        "type": "string"
                    },
                    "long_description": {
                        "description": "Long description of the category",
                        "type": "string"
                    },
                    "bom_id": {
                        "description": "ID of the bill of material associated with the category",
                        "type": "integer"
                    },
                    "updated_by": {
                        "description": "ID of the user who last updated the category",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}